+++ crates/nostr/examples/nip05.rs
@@ -10,7 +10,7 @@ fn main() -> Result<()> {
     let public_key =
         PublicKey::from_str("b2d670de53b27691c0c3400225b65c35a26d06093bcc41f48ffc71e0907f9d4a")?;
 
-    if nip05::verify_blocking(&public_key, "0xtr@oxtr.dev", None).is_ok() {
+    if nip05::verify_blocking(&public_key, "0xtr@oxtr.dev", None)? {
         println!("NIP-05 verified");
     } else {
         println!("NIP-05 NOT verified");